requirements.txt
python-magic
python-magic-bin; platform_system == 'Windows'
videohash
jsonpickle